+++ CHANGELOG.md
@@ -11,6 +11,7 @@ is handled internally, and as such things could potentially break.
 
 ### Added
 
+- Support for 1.21.5 - celeste, nobaboy
 - Fishing catch timer HUD element - celeste
 - Option to revert fishing treasure catch messages to the pre-Backwater Bayou format - celeste
 - Option to mute Banshee - nobaboy